♻️ refactor: 데이터 처리 로직 커스텀 훅 생성해서 분리
